compiler: new-solver: first cut of Bubble phase
Implement phase that propagates subtyping constraints where the upper bound
is a bottom element or the lower bound is a top element. For example:

```
?1 :< {x : U8 Present | ?2}
```

We know that the unification variable `?1` must have a field called `x`.
We also know that the field must be present in `?1`, because present is
the bottom of the subtyping hierarchy. So we can rewrite this constraint
by adding the following extra constraints:

```
?1 :< {x : U8 Present | ?2}

?1 :=: {x : ?3 Present | ?4}
?3 :< U8
```

We can do the same thing for Taken record fields on the lower bound of
subtyping relations, and conversely for variants (Taken on upper bound,
Present on lower bound).